fix worker.js script imports on firefox
The utils.js script is imported via the manifest for content scripts and in a script tag for popup/options scripts. As worker.js runs as a MV3 service worker in Chrome, we use importScripts(). But worker.js can only run as a background script on Firefox, so we use a MV2 background page with script tags instead. Plus an add-on ID is set to enable the storage API on Firefox.